# CRM_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Appra CRM API",
    "description": "Python/PostgreSQL CRM. No authentication. All business dates use Asia/Kolkata. All timestamps are returned in UTC. Only customers and follow-ups are stored. Validation errors return HTTP 400.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a customer and initial follow-up atomically",
        "description": "All strings are trimmed. source defaults to OTHER, status to NEW and preferredContact to WHATSAPP.\nIf either insert fails, neither record is saved. Phone is stored as a string.",
        "operationId": "create_customer_api_customers_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              },
              "example": {
                "name": "Arun Kumar",
                "phone": "+919876543210",
                "requirement": "Website Development",
                "source": "WEBSITE",
                "location": "Coimbatore",
                "notes": "Interested in company website",
                "followUp": {
                  "date": "2026-09-20",
                  "time": "10:30",
                  "preferredContact": "WHATSAPP"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CustomerDetail_"
                },
                "example": {
                  "data": {
                    "id": 8,
                    "name": "Arun Kumar",
                    "phone": "+919876543210",
                    "requirement": "Website Development",
                    "source": "WEBSITE",
                    "location": "Coimbatore",
                    "notes": "Interested in company website",
                    "status": "NEW",
                    "closedAt": null,
                    "createdAt": "2026-09-18T04:00:00Z",
                    "updatedAt": "2026-09-18T04:00:00Z",
                    "followUps": [
                      {
                        "id": 15,
                        "customerId": 8,
                        "customerName": "Arun Kumar",
                        "phone": "+919876543210",
                        "requirement": "Website Development",
                        "followUpAt": "2026-09-20T05:00:00Z",
                        "preferredContact": "WHATSAPP",
                        "status": "PENDING",
                        "notes": null,
                        "rescheduleCount": 0,
                        "completedAt": null,
                        "createdAt": "2026-09-18T04:00:00Z",
                        "updatedAt": "2026-09-18T04:00:00Z",
                        "isOverdue": false
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, query parameters, range, or ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid request. Check fields, enums, IDs and date ranges.",
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Customer or follow-up not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Customer or follow-up not found.",
                  "error": "Not Found"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update conflict; retry the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "The record changed concurrently. Please retry the operation.",
                  "error": "Conflict"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error; no private details exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 500,
                  "message": "An unexpected server error occurred.",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Search and filter customers with pagination",
        "description": "Sort: createdAt descending, then id descending. Search matches name, phone or requirement.",
        "operationId": "list_customers_api_customers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive literal substring search across name, phone and requirement.",
              "title": "Search"
            },
            "description": "Case-insensitive literal substring search across name, phone and requirement."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_CustomerOut_"
                },
                "example": {
                  "data": [
                    {
                      "id": 8,
                      "name": "Arun Kumar",
                      "phone": "+919876543210",
                      "requirement": "Website Development",
                      "source": "WEBSITE",
                      "location": "Coimbatore",
                      "notes": "Interested in company website",
                      "status": "NEW",
                      "closedAt": null,
                      "createdAt": "2026-09-18T04:00:00Z",
                      "updatedAt": "2026-09-18T04:00:00Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 20,
                    "total": 1,
                    "totalPages": 1
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, query parameters, range, or ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid request. Check fields, enums, IDs and date ranges.",
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Customer or follow-up not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Customer or follow-up not found.",
                  "error": "Not Found"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update conflict; retry the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "The record changed concurrently. Please retry the operation.",
                  "error": "Conflict"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error; no private details exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 500,
                  "message": "An unexpected server error occurred.",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/follow-ups/calendar": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "Get all calendar entries for an inclusive date range",
        "description": "Includes PENDING and COMPLETED follow-ups. Unpaginated, with a maximum 366-day range.",
        "operationId": "calendar_api_follow_ups_calendar_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Inclusive first business date. Maximum range 366 days.",
              "examples": [
                "2026-09-01"
              ],
              "title": "From"
            },
            "description": "Inclusive first business date. Maximum range 366 days."
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "examples": [
                "2026-09-30"
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_FollowUpOut__"
                },
                "example": {
                  "data": [
                    {
                      "id": 15,
                      "customerId": 8,
                      "customerName": "Arun Kumar",
                      "phone": "+919876543210",
                      "requirement": "Website Development",
                      "followUpAt": "2026-09-20T05:00:00Z",
                      "preferredContact": "WHATSAPP",
                      "status": "PENDING",
                      "notes": null,
                      "rescheduleCount": 0,
                      "completedAt": null,
                      "createdAt": "2026-09-18T04:00:00Z",
                      "updatedAt": "2026-09-18T04:00:00Z",
                      "isOverdue": false
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, query parameters, range, or ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid request. Check fields, enums, IDs and date ranges.",
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Customer or follow-up not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Customer or follow-up not found.",
                  "error": "Not Found"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update conflict; retry the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "The record changed concurrently. Please retry the operation.",
                  "error": "Conflict"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error; no private details exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 500,
                  "message": "An unexpected server error occurred.",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/follow-ups/notification-summary": {
      "get": {
        "tags": [
          "Follow-ups"
        ],
        "summary": "Get calculated today and overdue badge counts",
        "description": "total is today + overdue. These sets overlap for pending appointments earlier today.",
        "operationId": "notifications_api_follow_ups_notification_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_NotificationSummary_"
                },
                "example": {
                  "data": {
                    "today": 8,
                    "overdue": 3,
                    "total": 11
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, query parameters, range, or ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid request. Check fields, enums, IDs and date ranges.",
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Customer or follow-up not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Customer or follow-up not found.",
                  "error": "Not Found"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update conflict; retry the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "The record changed concurrently. Please retry the operation.",
                  "error": "Conflict"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error; no private details exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 500,
                  "message": "An unexpected server error occurred.",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/follow-ups/{id}/complete": {
      "patch": {
        "tags": [
          "Follow-ups"
        ],
        "summary": "Mark a follow-up complete",
        "description": "No body required. Repeated completion is idempotent and preserves completedAt.",
        "operationId": "complete_follow_up_api_follow_ups__id__complete_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9]\\d*$",
              "description": "Positive 32-bit integer ID.",
              "examples": [
                "1"
              ],
              "title": "Id"
            },
            "description": "Positive 32-bit integer ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_FollowUpOut_"
                },
                "example": {
                  "data": {
                    "id": 15,
                    "customerId": 8,
                    "customerName": "Arun Kumar",
                    "phone": "+919876543210",
                    "requirement": "Website Development",
                    "followUpAt": "2026-09-20T05:00:00Z",
                    "preferredContact": "WHATSAPP",
                    "status": "COMPLETED",
                    "notes": null,
                    "rescheduleCount": 0,
                    "completedAt": "2026-09-18T06:30:00Z",
                    "createdAt": "2026-09-18T04:00:00Z",
                    "updatedAt": "2026-09-18T04:00:00Z",
                    "isOverdue": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, query parameters, range, or ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid request. Check fields, enums, IDs and date ranges.",
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Customer or follow-up not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Customer or follow-up not found.",
                  "error": "Not Found"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update conflict; retry the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 409,
                  "message": "The record changed concurrently. Please retry the operation.",
                  "error": "Conflict"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error; no private details exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 500,
                  "message": "An unexpected server error occurred.",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Arun Kumar"
            ]
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 7,
            "title": "Phone",
            "description": "String; preserves country code and leading zeroes. 7-15 digits with optional +, spaces, parentheses, hyphens.",
            "examples": [
              "+919876543210"
            ]
          },
          "requirement": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Requirement",
            "examples": [
              "Website Development"
            ]
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource",
            "default": "OTHER",
            "examples": [
              "WEBSITE"
            ]
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "examples": [
              "Coimbatore"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "examples": [
              "Interested in company website"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus",
            "default": "NEW"
          },
          "followUp": {
            "$ref": "#/components/schemas/InitialFollowUp",
            "description": "Required initial follow-up; created in the same transaction as the customer."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "phone",
          "requirement",
          "followUp"
        ],
        "title": "CustomerCreate"
      },
...
      "LeadStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "CONTACTED",
          "INTERESTED",
          "FOLLOW_UP",
          "CLOSED",
          "NOT_INTERESTED"
        ],
        "title": "LeadStatus"
      },
...
    }
  }
}